Refactor 068.
